requirements.txt
numpy==1.26.1
xformers==0.0.27
Pillow
huggingface_hub
einops
safetensors
matplotlib
open3d
opencv-python
DateTime
plyfile
HTML4Vision
timm
tensorboardX
imgaug
iopath
imagecorruptions
loguru
pyhocon
numpy-quaternion==2022.4.3
trimesh==4.0.1
tqdm
pyrender
yapf>=0.40.2
gradio
pydantic==2.10.6
rerun-sdk==0.23.1
gradio-rerun==0.23.1
